add information handler and integrate helper and profile to it
