+++ lambda/main/src/eventHandler/Handlers/handlers.ts
@@ -1,15 +1,13 @@
 import { Client } from "@line/bot-sdk";
-import { HelperHandler } from "./help";
 import { TextMessageHandler } from "./textMessage";
-import { ProfileHandler } from "./profile";
 import { IMessageDb } from "database/dynamoDb/IMessageDb";
 import { ImageHandler } from "./image";
+import { InformationHandler } from "./informationHandler";
 
 export const handlersFactory = (messageDb: IMessageDb, lineClient: Client) => {
     return {
-        helperHandler: new HelperHandler(messageDb),
         textMessageHandler: new TextMessageHandler(messageDb),
-        profileHandler: new ProfileHandler(messageDb),
-        imageHandler: new ImageHandler(lineClient)
+        imageHandler: new ImageHandler(lineClient),
+        informationHandler: new InformationHandler(messageDb)
     }
 }
