+++ lambda/main/src/main.ts
@@ -4,9 +4,6 @@ import { HostLineClient, UserLineClient } from "eventHandler/lineClients";
 import { Handler } from "eventHandler/handler";
 import { handlersFactory } from "eventHandler/Handlers/handlers";
 import { MessageDb } from "database/dynamoDb/messageDb";
-import { richMenuObjectA } from "./richMenu/richMenu";
-import * as fs from "fs";
-import { join } from "path";
 
 export const lambdaHandler = async  (event: APIGatewayProxyEvent) => {
 
@@ -16,22 +13,15 @@ export const lambdaHandler = async  (event: APIGatewayProxyEvent) => {
 
     const hostClient = new HostLineClient();
 
+
     const handlers = handlersFactory(new MessageDb(), lineClient);
-
-    // const richMenu = await lineClient.createRichMenu(richMenuObjectA())
-    // const filePath = join(__dirname, "./16984988403378.png");
-    // const buffer = fs.readFileSync(filePath);
-
-    // await lineClient.setRichMenuImage(filePath, buffer);
-    // await lineClient.setDefaultRichMenu(richMenu);
 
     const handler = new Handler(
         lineClient,
         hostClient,
-        handlers.helperHandler,
-        handlers.profileHandler,
         handlers.textMessageHandler,
-        handlers.imageHandler
+        handlers.imageHandler,
+        handlers.informationHandler
     );
 
     const body: WebhookRequestBody = JSON.parse(event.body!);
@@ -41,7 +31,7 @@ export const lambdaHandler = async  (event: APIGatewayProxyEvent) => {
     } catch (error) {
 
         console.error(error);
-    
+
         if (error instanceof Error) {
             await hostClient.broadcast({
                 type: "text",
